backend/requirements.txt
# AI Strategy Hub — Backend Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
python-multipart==0.0.9

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.11.0
python-docx==1.1.0

# Vector Database & Embeddings
# (Switched to Gemini Embeddings + lightweight JSON storage to avoid chromadb C++ Build Tool issues)

# LLM Integration
google-generativeai==0.8.0

# Research Agent
tavily-python==0.5.0

# Database
aiosqlite==0.20.0

# Export
reportlab==4.2.0

# Security & Middleware
slowapi==0.1.9

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0

# Database Migrations
alembic==1.14.0

# Utilities
httpx==0.27.0
python-dotenv==1.0.1
